zyrabit-brain-api/api-rag/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-ollama    
chromadb-client     
prometheus-fastapi-instrumentator 
python-dotenv      
pytest
pytest-mock
httpx
